Feat: smooth zoom + continuous fractional pinch zoom
smooth-zoom: eased display-layer zoom animation for discrete zoom input
(buttons, wheel, keyboard) with geographic anchoring, retracking on rapid
input, and render-completion handoff (aligned immediate swap / 150ms
crossfade); composes with smooth-follow.

continuous-pinch-zoom: continuous fractional magnification from pinch
gestures - magnification is a Double end-to-end (JNI render/projection
boundary takes the 2^z scale factor via SetMagnification), pinch commits
unrounded startMag + log2(factor) with a 16x per-gesture cap, discrete
controls stay level-stepped, tile path composes floor-level tiles scaled
by 2^(z - floor(z)), gesture display scale folds/hands off without snaps
(base rescale invariant across buffer swaps), and viewport persistence
stores the fractional magnification (legacy integer files load as-is).

Includes input hardening (gestureStartDist gate + factor smoothing),
:auto renderer wired to viewportZoomFraction, ZoomAnimation/commit unit
tests, MapRendering.md fractional-mag rule, and both openspec changes
archived.

+++ app/src/main/java/com/naviveylin/data/ViewportState.kt
@@ -7,7 +7,8 @@ import kotlinx.serialization.Serializable
 data class ViewportState(
     val centerLat: Double = DEFAULT_LAT,
     val centerLon: Double = DEFAULT_LON,
-    val magnification: Int = DEFAULT_MAG,
+    /** Magnification scale factor (2^z, fractional z from continuous pinch). */
+    val magnification: Double = DEFAULT_MAG.toDouble(),
     /** Map rotation in radians, matching the native `MercatorProjection::Set` angle convention. */
     val angle: Double = DEFAULT_ANGLE
 ) {
